pyproject.toml
[project]
name = "metrics"
version = "0.1.0"
description = "Тестовое задание для Татнефть"
authors = [
    {name = "Илья Ватютов",email = "vatut007@ya.ru"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=6.0.2,<7.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "djangorestframework (>=3.16.1,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
